+++ test/e2e/btc_staking_pre_approval_e2e_test.go
@@ -492,14 +492,26 @@ func (s *BTCStakingPreApprovalTestSuite) Test5SubmitStakerUnbonding() {
 	s.NoError(err)
 	stakingTxHash := stakingMsgTx.TxHash()
 
-	// delegator signs unbonding tx
-	params := nonValidatorNode.QueryBTCStakingParams()
-	delUnbondingSig, err := activeDel.SignUnbondingTx(params, s.net, s.delBTCSK)
+	currentBtcTipResp, err := nonValidatorNode.QueryTip()
 	s.NoError(err)
+	currentBtcTip, err := chain.ParseBTCHeaderInfoResponseToInfo(currentBtcTipResp)
+	s.NoError(err)
+
+	unbondingTx := activeDel.BtcUndelegation.UnbondingTx
+	unbondingTxMsg, err := bbn.NewBTCTxFromBytes(unbondingTx)
+	s.NoError(err)
+
+	blockWithUnbondingTx := datagen.CreateBlockWithTransaction(s.r, currentBtcTip.Header.ToBlockHeader(), unbondingTxMsg)
+	nonValidatorNode.InsertHeader(&blockWithUnbondingTx.HeaderBytes)
+	inclusionProof := bstypes.NewInclusionProofFromSpvProof(blockWithUnbondingTx.SpvProof)
 
 	nonValidatorNode.SubmitRefundableTxWithAssertion(func() {
 		// submit the message for creating BTC undelegation
-		nonValidatorNode.BTCUndelegate(&stakingTxHash, delUnbondingSig)
+		nonValidatorNode.BTCUndelegate(
+			&stakingTxHash,
+			unbondingTxMsg,
+			inclusionProof,
+		)
 		// wait for a block so that above txs take effect
 		nonValidatorNode.WaitForNextBlock()
 	}, true)